Simplify time handling
Remove my 'date' computation because it turned out there's already
a 'time' entry which stamps when the prediction is for, and
convene on using unix epoch time instead of whatever datetime.date
prints itself as by default (which is apparently ISO 8601).

Also, hardcode the header order because that makes it tidier,
and I realized that by only reading headers off the first row
some data was being silently suppressed (e.g. precipAccum).
In my test data set I happened to get a day that had /most/ of
the headers because it was a rainy day, but that's not guaranteed.
So it's overall faster at this point to hardcode the headers,
especially because many columns are related so order does matter,
at least to us here humans.